// web/src/api/generated/models/user-word-book-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DailyWordResponse } from './daily-word-response';

/**
 * 用户生词本响应
 */
export interface UserWordBookResponse {
    /**
     * ID
     */
    'id'?: number;
    /**
     * 用户ID
     */
    'userId'?: number;
    /**
     * 单词ID
     */
    'wordId'?: number;
    /**
     * 学习状态
     */
    'learningStatus'?: number;
    /**
     * 学习状态描述
     */
    'learningStatusDesc'?: string;
    /**
     * 收藏时间
     */
    'collectedTime'?: string;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
    'word'?: DailyWordResponse;
}